# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is shopping server api server",
        "title": "Product api",
        "termsOfService": "not much usefull",
        "contact": {
            "name": "Asliddin",
            "url": "https://t.me/asliddindeh",
            "email": "asliddinvstalim@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8070",
    "basePath": "/v1",
    "paths": {
        "/product/delete/{id}": {
            "delete": {
                "description": "this func delete products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "delete products",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.GetProductId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/products": {
            "post": {
                "description": "this func create products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "create products",
                "parameters": [
                    {
                        "description": "Products",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.ProductResponse"
                        }
                    }
                }
            }
        },
        "/products/category": {
            "post": {
                "description": "this func create category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "create category",
                "parameters": [
                    {
                        "description": "Category",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.CategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.Category"
                        }
                    }
                }
            }
        },
        "/products/fullinfo": {
            "post": {
                "description": "this func create products with ful info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "create products with info",
                "parameters": [
                    {
                        "description": "Products",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.ProductFullInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.ProductFullInfoResponse"
                        }
                    }
                }
            }
        },
        "/products/info": {
            "get": {
                "description": "this func get all products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "get all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.Products"
                        }
                    }
                }
            }
        },
        "/products/type": {
            "post": {
                "description": "this func create type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "type"
                ],
                "summary": "create type",
                "parameters": [
                    {
                        "description": "Type",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.TypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.Type"
                        }
                    }
                }
            }
        },
        "/products/update": {
            "patch": {
                "description": "this func update product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "update product",
                "parameters": [
                    {
                        "description": "Products",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.Product"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "this func get products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "get products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/store": {
            "post": {
                "description": "this func create store with ful info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "create store",
                "parameters": [
                    {
                        "description": "Store",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.StoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.StoreResponse"
                        }
                    }
                }
            }
        },
        "/store/{id}": {
            "get": {
                "description": "this func get store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "get store",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    },
    "definitions": {
        "product.Address": {
            "type": "object",
            "properties": {
                "district": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "product.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "product.CategoryRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "product.GetProductId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "product.Product": {
            "type": "object",
            "properties": {
                "categoryid": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "typeid": {
                    "type": "integer"
                }
            }
        },
        "product.ProductFullInfo": {
            "type": "object",
            "properties": {
                "categoryid": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Store"
                    }
                },
                "typeid": {
                    "type": "integer"
                }
            }
        },
        "product.ProductFullInfoResponse": {
            "type": "object",
            "properties": {
                "categoryid": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Store"
                    }
                },
                "typeid": {
                    "type": "integer"
                }
            }
        },
        "product.ProductRequest": {
            "type": "object",
            "properties": {
                "categoryid": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "typeid": {
                    "type": "integer"
                }
            }
        },
        "product.ProductResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "product.Products": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.ProductResponse"
                    }
                }
            }
        },
        "product.Store": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Address"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "product.Type": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "product.TypeRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "store.Address": {
            "type": "object",
            "properties": {
                "district": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "store.AddressResp": {
            "type": "object",
            "properties": {
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "store.StoreRequest": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.Address"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "store.StoreResponse": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.AddressResp"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}